+++ src/shared_queue.cc
@@ -27,22 +27,67 @@ namespace proxy_wasm {
 
 SharedQueue global_shared_queue;
 
-SharedQueue::SharedQueue() {
-  registerVmIdHandleCallback([this](std::string_view vm_id) { this->deleteByVmId(vm_id); });
+SharedQueue::SharedQueue(bool register_vm_id_callback) {
+  if (register_vm_id_callback) {
+    registerVmIdHandleCallback([this](std::string_view vm_id) { this->deleteByVmId(vm_id); });
+  }
 }
 
-void SharedQueue::deleteByVmId(std::string_view vm_id) {}
+void SharedQueue::deleteByVmId(std::string_view vm_id) {
+  std::lock_guard<std::mutex> lock(mutex_);
+  auto queue_keys = vm_queue_keys_.find(std::string(vm_id));
+  if (queue_keys != vm_queue_keys_.end()) {
+    for (std::size_t queue_key : queue_keys->second) {
+      auto token = queue_tokens_.find(queue_key);
+      if (token != queue_tokens_.end()) {
+        queues_.erase(token->second);
+        queue_tokens_.erase(token);
+      }
+    }
+    vm_queue_keys_.erase(queue_keys);
+  }
+}
+
+uint32_t SharedQueue::nextQueueToken() {
+  // TODO(@mathetake): Handle the case where the queue overflows, i.e. the number of used
+  // tokens exceeds the max of uint32. If it overflows, the following loops never exits.
+  while (true) {
+    uint32_t token = next_queue_token_++;
+    if (token == 0) {
+      continue; // 0 is an illegal token.
+    }
+
+    if (queues_.find(token) == queues_.end()) {
+      return token;
+    }
+  }
+}
 
 uint32_t SharedQueue::registerQueue(std::string_view vm_id, std::string_view queue_name,
                                     uint32_t context_id, CallOnThreadFunction call_on_thread,
                                     std::string_view vm_key) {
   std::lock_guard<std::mutex> lock(mutex_);
-  auto key = std::make_pair(std::string(vm_id), std::string(queue_name));
+
+  // TODO(mathetake): care about hash collision?
+  auto key = queueKeyHash(vm_id, queue_name);
+  std::unordered_set<std::size_t> *queue_keys;
+  std::string vid = std::string(vm_id);
+  auto map_it = vm_queue_keys_.find(vid);
+  if (map_it == vm_queue_keys_.end()) {
+    queue_keys = &vm_queue_keys_[vid];
+  } else {
+    queue_keys = &map_it->second;
+  }
+  queue_keys->insert(key);
+
   auto it = queue_tokens_.insert(std::make_pair(key, static_cast<uint32_t>(0)));
   if (it.second) {
     it.first->second = nextQueueToken();
-    queue_token_set_.insert(it.first->second);
+    if (it.first->second == 0) {
+      return 0; // queue overflows
+    }
   }
+
   uint32_t token = it.first->second;
   auto &q = queues_[token];
   q.vm_key = std::string(vm_key);
@@ -54,7 +99,7 @@ uint32_t SharedQueue::registerQueue(std::string_view vm_id, std::string_view que
 
 uint32_t SharedQueue::resolveQueue(std::string_view vm_id, std::string_view queue_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  auto key = std::make_pair(std::string(vm_id), std::string(queue_name));
+  auto key = queueKeyHash(vm_id, queue_name);
   auto it = queue_tokens_.find(key);
   if (it != queue_tokens_.end()) {
     return it->second;
